+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.11.4
+* adds `queueName` as first argument to the `connector.startWorker` method
+* adds `queueName` as option to the `ctx.enqueue` method
+
 # 0.11.3
 * add `timeout` and `retry` to client api calls options; when defined thogether timeout specifies number of miliseconds after which the connection would be timeouted and retry specifies number of miliseconds when the retry will be done after the timeout
 * by default `client.get` method retries twice calls when error 503 is returned